+++ src/deepskylog/AstronomyLibrary/Coordinates/EquatorialCoordinates.php
@@ -14,6 +14,7 @@
 namespace deepskylog\AstronomyLibrary\Coordinates;
 
 use Carbon\Carbon;
+use deepskylog\AstronomyLibrary\Time;
 
 /**
  * EquatorialCoordinates class.
@@ -548,7 +549,7 @@ class EquatorialCoordinates
      */
     public function precession(Carbon $date): EquatorialCoordinates
     {
-        $precessed_coordinates = $this;
+        $precessed_coordinates = clone $this;
 
         if ($date->isLeapYear()) {
             $year = $date->year + ($date->dayOfYear - 1.0) / 366;
@@ -581,5 +582,57 @@ class EquatorialCoordinates
 
         return $precessed_coordinates;
     }
+
+    /**
+     * Returns the precession: the coordinates for another epoch and equinox.
+     * Chapter 21 of Astronomical Algorithms.
+     *
+     * @param Carbon $date The date for the new equinox
+     *
+     * @return EquatorialCoordinates the diameter of the smallest circle
+     */
+    public function precessionHighAccuracy(Carbon $date): EquatorialCoordinates
+    {
+        $precessed_coordinates = clone $this;
+
+        $epoch_in_JD = Time::getJd(
+            Carbon::create($this->getEpoch(), 1, 1, 12, 0, 0, 'UTC')
+        );
+
+        $time_interval_J2000_starting = ($epoch_in_JD - 2451545.0) / 36525.0;
+
+        $jd = Time::getJd($date);
+
+        $time_interval_starting_final = ($jd - $epoch_in_JD) / 36525.0;
+
+        $ra_with_proper_motion = (
+            $this->getRA()->getCoordinate()
+            + $this->getDeltaRA() * $time_interval_starting_final * 100.0 / 3600.0
+        ) * 15.0;
+        $dec_with_proper_motion = $this->getDeclination()->getCoordinate()
+            + $this->getDeltaDec() * $time_interval_starting_final * 100.0 / 3600.0;
+
+        $ksi = ((2306.2181 + 1.39656 * $time_interval_J2000_starting - 0.000139 * $time_interval_J2000_starting ** 2) * $time_interval_starting_final
+            + (0.30188 - 0.000344 * $time_interval_J2000_starting) * $time_interval_starting_final ** 2
+            + 0.017998 * $time_interval_starting_final ** 3) / 3600.0;
+
+        $zeta = ((2306.2181 + 1.39656 * $time_interval_J2000_starting - 0.000139 * $time_interval_J2000_starting ** 2) * $time_interval_starting_final
+            + (1.09468 + 0.000066 * $time_interval_J2000_starting) * $time_interval_starting_final ** 2
+            + 0.018203 * $time_interval_starting_final ** 3) / 3600.0;
+
+        $theta = ((2004.3109 - 0.85330 * $time_interval_J2000_starting - 0.000217 * $time_interval_J2000_starting ** 2) * $time_interval_starting_final
+            - (0.42665 + 0.000217 * $time_interval_J2000_starting) * $time_interval_starting_final ** 2
+            - 0.041833 * $time_interval_starting_final ** 3) / 3600.0;
+
+        $A = cos(deg2rad($dec_with_proper_motion)) * sin(deg2rad($ra_with_proper_motion + $ksi));
+        $B = cos(deg2rad($theta)) * cos(deg2rad($dec_with_proper_motion)) * cos(deg2rad($ra_with_proper_motion + $ksi)) - sin(deg2rad($theta)) * sin(deg2rad($dec_with_proper_motion));
+        $C = sin(deg2rad($theta)) * cos(deg2rad($dec_with_proper_motion)) * cos(deg2rad($ra_with_proper_motion + $ksi)) + cos(deg2rad($theta)) * sin(deg2rad($dec_with_proper_motion));
+
+        $precessed_coordinates->setRA(
+            (rad2deg(atan2($A, $B)) + $zeta) / 15.0
+        );
+        $precessed_coordinates->setDeclination(rad2deg(asin($C)));
+
+        return $precessed_coordinates;
     }
 }